Bugfix: cd back into the top-level directory when there's a failure.
When the code was returning early due to a failure, it was staying in a
temporary subdirectory used for testing.  This is obviously wrong as it
confused the rest of the configure script which assumed it was still running
from the top-level directory, and would cause errors such as:
  config.status: error: cannot find input file: build-aux/Makefile.in

Bug report and fix provided by Guillaume Lazzara.